requirements.txt
openai
langchain
vllm
transformers
google-genai
replicate